refactor(editor): meta model 분리와 code fallback 복구
